feat(hashtable): split into two files
